Refactor org-detail
