+++ components/analytics/org-detail/index.tsx
@@ -1,0 +1,1 @@
+export { OrgDetail } from "./org-detail";
